Adjust cpu_settings and rename
Renaming because there might be other system related stuff here and it
would be convenient to use the same file.

This also rips out the support for 32 bit systems and mips architectures
because we are a small team and we currently don't want to over
complicate things. Additionally pointer business in Crystal is a bit
tricky so locking ourselves to only 64 bit pointers will simplify things
until later.